implement AsAny only for Request type
